+++ packages/date/stories/date.stories.tsx
@@ -74,7 +74,7 @@ export const DateInput: Story = ({ datepicker, disabled, max, min, required }) =
 };
 DateInput.storyName = 'date';
 
-export const _DateField: Story = ({ datepicker, disabled, label, max, min, required }) => {
+export const _DateField: Story = ({ datepicker, disabled, helpId, label, max, min, required }) => {
   const dateFormat = 'MM/DD/YYYY';
   const minDate = moment(min).format(dateFormat);
   const maxDate = moment(max).format(dateFormat);
@@ -102,6 +102,8 @@ export const _DateField: Story = ({ datepicker, disabled, label, max, min, requi
         id="dateOfService"
         name="dateOfService"
         label={label}
+        required={required}
+        helpId={helpId}
         disabled={disabled}
         datepicker={datepicker}
         min={min && minDate}
@@ -116,6 +118,7 @@ export const _DateField: Story = ({ datepicker, disabled, label, max, min, requi
 };
 _DateField.args = {
   label: 'Date Field',
+  helpId: '',
 };
 _DateField.storyName = 'DateField';
 
@@ -205,6 +208,7 @@ export const _DateRangeField: Story = ({
   autoSync,
   datepicker,
   disabled,
+  helpId,
   max,
   maxDistance,
   maxDistanceUnits,
@@ -254,6 +258,8 @@ export const _DateRangeField: Story = ({
         name="dateOfService"
         label="Date of Service"
         disabled={disabled}
+        required={required}
+        helpId={helpId}
         datepicker={datepicker}
         autoSync={autoSync}
         ranges={ranges}
@@ -269,6 +275,7 @@ export const _DateRangeField: Story = ({
 };
 _DateRangeField.args = {
   autoSync: false,
+  helpId: '',
   maxDistance: 0,
   maxDistanceUnits: distanceUnits[0],
   minDistance: 0,