+++ packages/form/src/CheckboxGroup.js
@@ -1,11 +1,11 @@
 import React, { createContext, useContext } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { FieldHelpIcon } from '@availity/help';
 import { useField, useFormikContext } from 'formik';
 
 import Feedback from './Feedback';
 import FormGroup from './FormGroup';
+import Label from './Label';
 
 export const CheckboxGroupContext = createContext();
 
@@ -34,7 +34,17 @@ export const useCheckboxGroup = (name) => {
   return { toggle, value: value.indexOf(name) > -1, ...rest };
 };
 
-const CheckboxGroup = ({ name, children, onChange: groupOnChange, groupClassName, label, helpId, ...rest }) => {
+const CheckboxGroup = ({
+  name,
+  children,
+  onChange: groupOnChange,
+  groupClassName,
+  label,
+  labelClassName,
+  required,
+  helpId,
+  ...rest
+}) => {
   const [field, metadata] = useField(name);
 
   const classes = classNames(
@@ -49,23 +59,24 @@ const CheckboxGroup = ({ name, children, onChange: groupOnChange, groupClassName
 
   if (label) {
     tag = 'fieldset';
-    legend = <legend className="text-dark">{label}</legend>;
+    const legendId = `${name}-legend`.toLowerCase();
+    const srRequiredAsterisk = required ? '* ' : null;
+    const styles = { cursor: 'default', lineHeight: 'inherit', color: '#000' };
+    const labelClasses = classNames('form-inline', labelClassName, !labelClassName && 'h4 font-weight-normal');
 
-    if (helpId) {
-      const legendId = `${name}-legend`.toLowerCase();
-      const legendStyle = { fontSize: '1.5rem', marginBottom: '.5rem' };
-      legend = (
-        <>
-          <legend id={legendId} className="sr-only">
+    legend = (
+      <>
+        <legend id={legendId} className="sr-only">
+          {srRequiredAsterisk}
+          {label}
+        </legend>
+        <div className={labelClasses} style={styles}>
+          <Label tag="div" aria-hidden helpId={helpId} required={required}>
             {label}
-          </legend>
-          <div className="form-inline text-dark" style={legendStyle}>
-            <div aria-hidden="true">{label}</div>
-            <FieldHelpIcon id={helpId} labelId={legendId} />
-          </div>
-        </>
-      );
-    }
+          </Label>
+        </div>
+      </>
+    );
   }
 
   return (
@@ -88,6 +99,8 @@ CheckboxGroup.propTypes = {
   label: PropTypes.node,
   name: PropTypes.string.isRequired,
   onChange: PropTypes.func,
+  labelClassName: PropTypes.string,
+  required: PropTypes.bool,
 };
 
 export default CheckboxGroup;