+++ packages/form/stories/form.stories.tsx
@@ -1,5 +1,4 @@
 /* eslint-disable import/no-duplicates */
-/* eslint-disable no-console */
 import React from 'react';
 import { Meta, Story } from '@storybook/react';
 import { Button } from 'reactstrap';
@@ -10,7 +9,7 @@ import { Phone } from '@availity/phone';
 import { DateField } from '@availity/date';
 import { SelectField } from '@availity/select';
 
-import { Field, Input, Checkbox, CheckboxGroup, RadioGroup, Radio, FormGroup } from '..';
+import { Field, Input, Checkbox, CheckboxGroup, Label, RadioGroup, Radio, FormGroup, RequiredKey } from '..';
 // import README from '../form/README.md';
 
 import FormResults from '../../../story-utils/FormikResults';
@@ -31,7 +30,7 @@ export const Default: Story = ({ required }) => {
   const schema = yup.object().shape({
     field: yup.string().isRequired(required, 'This field is required.'),
     phone: yup.string().validatePhone(undefined, false, 'US').isRequired(required, 'This field is required.'),
-    ext: yup.string().isRequired(required),
+    ext: yup.string(),
     dateField: avDate({
       format: 'MM/DD/YYYY',
     })
@@ -41,6 +40,7 @@ export const Default: Story = ({ required }) => {
       .between('', ''),
     checkboxGroup: yup.array().isRequired(required, 'At least one checkbox is required'),
     radioGroup: yup.string().isRequired(required, 'A selection is required'),
+    radioGroupLabelStyling: yup.string().isRequired(required, 'A selection is required'),
     SelectField: yup.string().isRequired(required, 'This field is required.').nullable(),
     disabledField: yup.string(),
     disabledDateField: avDate({
@@ -70,6 +70,7 @@ export const Default: Story = ({ required }) => {
         dateField: '',
         checkboxGroup: [],
         radioGroup: '',
+        radioGroupLabelStyling: '',
         SelectField: undefined,
         disabledField: '',
         disabledDateField: '01/01/2021',
@@ -78,29 +79,42 @@ export const Default: Story = ({ required }) => {
       validationSchema={schema}
     >
       <h2>Form</h2>
-
-      <Field name="field" type="text" label="Field" helpId="fieldHelp" />
+      <RequiredKey />
+      <br />
+      <Field name="field" type="text" label="Field" required={required} />
       <Phone
         name="phone"
         label="Phone"
         showExtension
+        required={required}
         extProps={{
           name: 'ext',
           label: 'Ext.',
         }}
       />
-      <DateField id="dateField" name="dateField" label="Date Field" datepicker={false} />
-      <CheckboxGroup name="checkboxGroup" label="Checkbox Group" helpId="checkboxGroup">
+      <DateField id="dateField" name="dateField" label="Date Field" required={required} datepicker={false} />
+      <CheckboxGroup name="checkboxGroup" label="Checkbox Group" helpId="checkboxGroup" required={required}>
         <Checkbox groupName="checkboxGroup" label="Check One" value="uno" helpId="option1" />
         <Checkbox groupName="checkboxGroup" label="Check Two" value="dos" />
         <Checkbox groupName="checkboxGroup" label="Check Three" value="tres" helpId="option3" />
       </CheckboxGroup>
-      <RadioGroup name="radioGroup" label="Radio Group">
+      <RadioGroup name="radioGroup" label="Radio Group" required={required}>
         <Radio name="radioGroup" label="Radio One" value="uno" />
         <Radio name="radioGroup" label="Radio Two" value="dos" />
         <Radio name="radioGroup" label="Radio Three" value="tres" />
       </RadioGroup>
-      <SelectField label="Select Field" name="SelectField" options={options} />
+      <RadioGroup
+        name="radioGroupLabelStyling"
+        label="Radio Group with Legend styled like Label"
+        labelClassName="label"
+        helpId="radioGroup"
+        required={required}
+      >
+        <Radio name="radioGroupLabelStyling" label="Radio One" value="uno" />
+        <Radio name="radioGroupLabelStyling" label="Radio Two" value="dos" />
+        <Radio name="radioGroupLabelStyling" label="Radio Three" value="tres" />
+      </RadioGroup>
+      <SelectField label="Select Field" name="SelectField" options={options} required={required} />
       <Field name="disabledField" type="text" label="Disabled Field" disabled />
       <DateField
         id="disabledDateField"
@@ -140,6 +154,18 @@ export const _Input: Story = ({ required }) => (
 );
 _Input.storyName = 'Input';
 
+export const _Label: Story = ({ required, helpId, className, children }) => (
+  <Label required={required} helpId={helpId} className={className}>
+    {children}
+  </Label>
+);
+_Label.args = {
+  helpId: '',
+  className: '',
+  children: 'Label Text',
+};
+_Label.storyName = 'Label';
+
 export const _FormGroup: Story = ({ required }) => (
   <FormResults
     onSubmit={() => {
@@ -172,7 +198,7 @@ export const _Field: Story = ({ required, helpMessage, helpId }) => (
       hello: yup.string().isRequired(required, 'This field is required.'),
     })}
   >
-    <Field name="hello" type="text" label="Hello" helpMessage={helpMessage} helpId={helpId} />
+    <Field name="hello" type="text" label="Hello" helpMessage={helpMessage} helpId={helpId} required={required} />
     <Button className="ml-1" color="primary" type="submit">
       Submit
     </Button>
@@ -184,7 +210,7 @@ _Field.args = {
 };
 _Field.storyName = 'Field';
 
-export const _Checkbox: Story = ({ required, helpId }) => (
+export const _Checkbox: Story = ({ required, helpId, labelClassName }) => (
   <FormResults
     onSubmit={() => {
       console.log('submitted');
@@ -196,7 +222,13 @@ export const _Checkbox: Story = ({ required, helpId }) => (
       checkboxGroup: yup.array().isRequired(required, 'At least one checkbox is required'),
     })}
   >
-    <CheckboxGroup name="checkboxGroup" helpId={helpId} label="Checkbox Group">
+    <CheckboxGroup
+      name="checkboxGroup"
+      helpId={helpId}
+      label="Checkbox Group"
+      required={required}
+      labelClassName={labelClassName}
+    >
       <Checkbox groupName="checkboxGroup" label="Check One" value="uno" />
       <Checkbox groupName="checkboxGroup" label="Check Two" value="dos" />
       <Checkbox groupName="checkboxGroup" label="Check Three" value="tres" />
@@ -208,10 +240,11 @@ export const _Checkbox: Story = ({ required, helpId }) => (
 );
 _Checkbox.args = {
   helpId: '',
+  labelClassName: 'label',
 };
 _Checkbox.storyName = 'Checkbox';
 
-export const _Radio: Story = ({ required, helpId }) => (
+export const _Radio: Story = ({ required, helpId, labelClassName }) => (
   <FormResults
     onSubmit={() => {
       console.log('submitted');
@@ -223,9 +256,9 @@ export const _Radio: Story = ({ required, helpId }) => (
       hello: yup.string().isRequired(required, 'This field is required.'),
     })}
   >
-    <RadioGroup name="hello" label="Radio Group" helpId={helpId}>
+    <RadioGroup name="hello" label="Radio Group" helpId={helpId} required={required} labelClassName={labelClassName}>
       <Radio name="hello" label="Radio One" value="uno" />
-      <Radio name="hello" label="Radio Two" value="dos" />
+      <Radio name="hello" label="Radio Two" value="dos" helpId="radioTwoHelp" />
       <Radio name="hello" label="Radio Three" value="tres" />
     </RadioGroup>
     <Button className="ml-1" color="primary" type="submit">
@@ -235,5 +268,6 @@ export const _Radio: Story = ({ required, helpId }) => (
 );
 _Radio.args = {
   helpId: '',
+  labelClassName: 'label',
 };
 _Radio.storyName = 'Radio';