+++ packages/reactstrap-validation-select/src/AvSelect.js
@@ -339,8 +339,8 @@ class AvSelect extends AvBaseInput {
           boxShadow: 0,
           colors: {
             ...theme.colors,
-            primary25: '#85a8dc',
-            primary: 'rgb(50 98 175)',
+            primary25: '#b8d4fb',
+            primary: '#85a8fc',
           },
         })}
         options={!attributes.loadOptions ? [...options, ...newOptions] : undefined}