+++ packages/select/src/SelectField.js
@@ -1,34 +1,21 @@
 import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { FormGroup, Feedback } from '@availity/form';
-import { FieldHelpIcon } from '@availity/help';
-import { Label as RsLabel } from 'reactstrap';
+import { FormGroup, Feedback, Label } from '@availity/form';
 
 import Select from './Select';
 
-const Label = ({ name, className, helpId, hidden, label }) => {
-  if (!label) return null;
-
-  return (
-    <>
-      <RsLabel id={`${name}-label`} for={name} hidden={hidden} className={className}>
-        {label}
-      </RsLabel>
-      {helpId ? <FieldHelpIcon labelId={`${name}-label`} id={helpId} /> : null}
-    </>
-  );
-};
-
-Label.propTypes = {
-  name: PropTypes.string.isRequired,
-  className: PropTypes.string,
-  helpId: PropTypes.string,
-  hidden: PropTypes.bool,
-  label: PropTypes.node,
-};
-
-const SelectField = ({ feedbackClass, groupClass, label, labelClass, labelHidden, name, helpId, ...attributes }) => {
+const SelectField = ({
+  feedbackClass,
+  groupClass,
+  label,
+  labelClass,
+  labelHidden,
+  name,
+  helpId,
+  required,
+  ...attributes
+}) => {
   useEffect(() => {
     if (attributes.id && attributes.id === name) {
       // eslint-disable-next-line no-console
@@ -38,10 +25,23 @@ const SelectField = ({ feedbackClass, groupClass, label, labelClass, labelHidden
     }
   }, [attributes.id, name]);
 
+  const thisLabel = label ? (
+    <Label
+      id={`${name}-label`}
+      for={name}
+      hidden={labelHidden}
+      className={labelClass}
+      required={required}
+      helpId={helpId}
+    >
+      {label}
+    </Label>
+  ) : null;
+
   return (
     <FormGroup className={groupClass} for={name} disabled={attributes.disabled}>
-      <Label name={name} label={label} hidden={labelHidden} className={labelClass} helpId={helpId} />
-      <Select name={name} {...attributes} />
+      {thisLabel}
+      <Select name={name} feedback required={required} {...attributes} />
       <Feedback className={classNames('d-block', feedbackClass)} name={name} />
     </FormGroup>
   );
@@ -55,6 +55,7 @@ SelectField.propTypes = {
   label: PropTypes.node,
   labelClass: PropTypes.string,
   labelHidden: PropTypes.bool,
+  required: PropTypes.bool,
 };
 
 export default SelectField;