+++ packages/select/stories/select.stories.tsx
@@ -1,11 +1,11 @@
 /* eslint-disable no-console */
 import React from 'react';
 import { Meta, Story } from '@storybook/react';
-import { Label, Button } from 'reactstrap';
+import { Button } from 'reactstrap';
 import AvApi from '@availity/api-axios';
 import * as yup from 'yup';
 import '@availity/yup';
-import { Feedback, FormGroup, Field } from '@availity/form';
+import { Feedback, FormGroup, Field, Label } from '@availity/form';
 
 import Select, { SelectField, ResourceSelect } from '..';
 // import README from '../README.md';
@@ -104,7 +104,6 @@ export const Default: Story = ({ autofill, creatable, disabled, isMulti, max, mi
       isMulti={isMulti}
       creatable={creatable}
       options={autofill ? autofillOptions : options}
-      minLength={min}
       maxLength={max}
       aria-label="stand-alone"
       raw={raw}
@@ -150,17 +149,16 @@ export const WithLabel: Story = ({
     }
   >
     <FormGroup for="standAloneWithLabel">
-      <Label for="standAloneWithLabel">{label}</Label>
+      <Label for="standAloneWithLabel" required={required}>
+        {label}
+      </Label>
       <Select
         name="standAloneWithLabel"
         autofill={autofill}
-        minLength={min}
         maxLength={max}
         isMulti={isMulti}
         options={autofill ? autofillOptions : options}
         creatable={creatable}
-        inputProps={{ 'aria-label': 'stand-alone with Label' }}
-        required={required}
         raw={raw}
         isDisabled={disabled}
       />
@@ -209,7 +207,6 @@ export const _SelectField: Story = ({
       name="SelectField"
       autofill={autofill}
       label={label}
-      minLength={min}
       maxLength={max}
       creatable={creatable}
       isMulti={isMulti}
@@ -246,12 +243,7 @@ export const _ResourceSelect: Story = ({ creatable, disabled, isMulti, label, ma
     }
   >
     <ResourceSelect
-      label={
-        <>
-          {label}
-          {required ? <span className="text-danger">*</span> : null}
-        </>
-      }
+      label={label}
       name="ResourceSelect"
       labelKey="name"
       maxLength={max}
@@ -285,12 +277,7 @@ export const GraphQlResourceSelect: Story = ({ creatable, disabled, isMulti, lab
     }
   >
     <ResourceSelect
-      label={
-        <>
-          {label}
-          {required ? <span className="text-danger">*</span> : null}
-        </>
-      }
+      label={label}
       name="ResourceSelect"
       labelKey="value"
       valueKey="id"