+++ packages/page-header/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.0.30](https://github.com/Availity/availity-react/compare/@availity/page-header@11.0.29...@availity/page-header@11.0.30) (2021-11-12)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
+## [11.0.29](https://github.com/Availity/availity-react/compare/@availity/page-header@11.0.27...@availity/page-header@11.0.29) (2021-11-12)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
+## [11.0.28](https://github.com/Availity/availity-react/compare/@availity/page-header@11.0.27...@availity/page-header@11.0.28) (2021-11-09)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
 ## [11.0.27](https://github.com/Availity/availity-react/compare/@availity/page-header@11.0.23...@availity/page-header@11.0.27) (2021-11-05)
 
 **Note:** Version bump only for package @availity/page-header